Refactored IntakeNote to Intake.
