+++ src/main/java/frc/robot/subsystems/intake/IntakeConstants.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.IntakeNote;
+package frc.robot.subsystems.intake;
 
 import frc.robot.Constants;
 import frc.robot.constants.controls.GearRatio;
