+++ src/main/java/frc/robot/subsystems/intake/IntakeNoteIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.IntakeNote;
+package frc.robot.subsystems.intake;
 
 import org.littletonrobotics.junction.AutoLog;
