+++ src/main/java/frc/robot/subsystems/intake/IntakeNoteIOPhysical.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.IntakeNote;
+package frc.robot.subsystems.intake;
 
 import com.revrobotics.CANSparkBase.IdleMode;
 import com.revrobotics.CANSparkLowLevel.MotorType;
